feat(ui-audit): Phase 1–6 UI audit — findings reports, component extraction, accessibility fixes (#4596)
* Initial plan

* Merge speckit.specify changes into copilot/sub-pr-4595

* Initial plan

* Add spec.md for UI Audit feature specification


---------


* feat(ui-audit): phases 1-6 — findings reports, component extraction, accessibility fixes

Phase 1 (T001-T005): Setup complete
- Created specs/052-ui-audit/findings/ directory
- Ran line count, lint, TypeScript, and test assessments (1145 tests passing)

Phase 2 (T006-T016): All 11 findings reports produced
- AppsPage, ProjectsPage, AgentsPipelinePage, AgentsPage, ChoresPage
- HelpPage, SettingsPage, ToolsPage, AppPage, LoginPage, NotFoundPage

Phase 6 (US4 - Modular Structure):
- AppsPage: extracted CreateAppDialog (707→223 lines), replaced zinc/emerald hardcoded colors with design tokens, added aria-pressed, aria-expanded, autoFocus
- ProjectsPage: extracted PipelineStagesPanel (board/PipelineStagesPanel.tsx) and BoardStatusBanners (board/BoardStatusBanners.tsx), removed IIFE in JSX, (629→385 lines)
- AgentsPipelinePage: removed console.warn

Phase 3 (US1): SettingsPage now shows error state when useUserSettings fails
Phase 4 (US2): a11y fixes across NotFoundPage, LoginPage, AppPage, HelpPage (aria-hidden, table caption, aria-pressed/expanded)
Phase 5 (US3): ChoresPage unsaved-changes modal replaced with shared ConfirmationDialog
Phase 9 (US7): Removed console.warn from AgentsPipelinePage and ChoresPage

---------
