+++ specs/052-ui-audit/findings/AppPage.md
@@ -1,0 +1,103 @@
+# Findings Report: AppPage
+
+**Page**: AppPage.tsx | **Route**: /
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 141 (limit: 250) ✅ Within limit
+**Feature Dir**: src/components/ (uses navigation only)
+**Related Hooks**: useNavigate (router only)
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 6 | 0 | 0 | 6 |
+| Loading/Error/Empty States | 4 | 1 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 5 | 2 | 0 | 7 |
+| Copy & UX Polish | 8 | 0 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 3 | 2 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **55** | **5** | **0** | **60** |
+
+**Overall**: 55/60 items passing
+
+## Findings
+
+### Major: STATE-03 — Static page has no applicable loading/error states
+
+**Dimension**: Loading/Error/Empty States
+**Checklist Item**: STATE-01/02/03 — Loading/Error/Empty
+**Location**: `src/pages/AppPage.tsx`
+**Related FR**: FR-004, FR-005, FR-006
+
+**Issue**: AppPage is a static landing/welcome page with no data fetching. All STATE checks are N/A. ✅
+
+---
+
+### Minor: A11Y-01 — Quick link buttons missing accessible names
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-01 — Keyboard accessible
+**Location**: `src/pages/AppPage.tsx:130-145`
+**Related FR**: FR-008
+
+**Issue**: Quick link `<button>` elements have visible text labels but no `aria-label` attribute. The labels are visible text content, which is sufficient. ✅ (This is a pass.)
+
+---
+
+### Minor: A11Y-07 — Decorative icons in quick links not aria-hidden
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-07 — Screen reader text
+**Location**: `src/pages/AppPage.tsx:130-145`
+**Related FR**: —
+
+**Issue**: The `link.icon` components inside quick link buttons are rendered without `aria-hidden="true"`. Since the button has visible text, the icon is decorative and should be hidden from screen readers.
+
+**Remediation**:
+Add `aria-hidden="true"` to the icon component: `<link.icon className="h-5 w-5" aria-hidden="true" />`.
+
+---
+
+### Minor: A11Y-03 — Decorative orbital/halo elements not aria-hidden
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-03 — ARIA attributes
+**Location**: `src/pages/AppPage.tsx:78-100`
+**Related FR**: FR-010
+
+**Issue**: Decorative UI elements (celestial-orbit, solar-halo, lunar-disc, hanging-stars) are purely visual but not explicitly hidden from screen readers.
+
+**Remediation**:
+Add `aria-hidden="true"` to the decorative container div.
+
+---
+
+### Minor: TEST-02 — AppPage has no component test for navigation
+
+**Dimension**: Test Coverage
+**Checklist Item**: TEST-02 — Component tests
+**Location**: `src/pages/AppPage.test.tsx` (check if exists)
+**Related FR**: FR-032
+
+**Issue**: No test verifying quick link navigation behavior.
+
+---
+
+## Remediation Plan
+
+### Phase 1: Accessibility
+- [ ] A11Y-07 — Add aria-hidden="true" to quick link icons
+- [ ] A11Y-03 — Add aria-hidden="true" to decorative container
+
+### Phase 2: Testing
+- [ ] TEST-02 — Add navigation tests for quick links
+
+### Phase 3: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
