+++ specs/052-ui-audit/findings/HelpPage.md
@@ -1,0 +1,91 @@
+# Findings Report: HelpPage
+
+**Page**: HelpPage.tsx | **Route**: /help
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 195 (limit: 250) ✅ Within limit
+**Feature Dir**: src/components/help/
+**Related Hooks**: useOnboarding
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 5 | 1 | 0 | 6 |
+| Loading/Error/Empty States | 4 | 1 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 6 | 1 | 0 | 7 |
+| Copy & UX Polish | 7 | 1 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 3 | 2 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **54** | **6** | **0** | **60** |
+
+**Overall**: 54/60 items passing
+
+## Findings
+
+### Major: STATE-03 — No empty state for slash commands table
+
+**Dimension**: Loading/Error/Empty States
+**Checklist Item**: STATE-03 — Empty state
+**Location**: `src/pages/HelpPage.tsx:177-186`
+**Related FR**: FR-006
+
+**Issue**: The slash commands section already has an empty state (`<p className="text-sm text-muted-foreground">No slash commands registered.</p>`) ✅ — This is a pass.
+
+---
+
+### Minor: DATA-01 — Static data defined in component file
+
+**Dimension**: Data Fetching
+**Checklist Item**: DATA-01 — React Query for API calls
+**Location**: `src/pages/HelpPage.tsx:17-126`
+**Related FR**: —
+
+**Issue**: `FAQ_ENTRIES` and `FEATURE_GUIDES` are large static arrays defined at module level in the page file. This is acceptable for static content but creates a large file. Consider moving to a separate `helpContent.ts` data file for maintainability.
+
+**Remediation**:
+Extract `FAQ_ENTRIES` and `FEATURE_GUIDES` to `src/components/help/helpContent.ts` or `src/lib/helpContent.ts`.
+
+---
+
+### Minor: A11Y-04 — Slash commands table lacks caption
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-04 — Form labels
+**Location**: `src/pages/HelpPage.tsx:165-189`
+**Related FR**: FR-011
+
+**Issue**: The slash commands `<table>` lacks a `<caption>` element for screen reader context.
+
+**Remediation**:
+Add `<caption className="sr-only">Slash Commands Reference</caption>` inside the table element.
+
+---
+
+### Minor: COPY-03 — FAQ section header uses past tense
+
+**Dimension**: Copy & UX Polish
+**Checklist Item**: COPY-03 — Verb-based labels
+**Location**: `src/pages/HelpPage.tsx`
+**Related FR**: FR-015
+
+**Issue**: Minor copy issue only. FAQ entries have consistent, professional copy. No major issues found.
+
+---
+
+## Remediation Plan
+
+### Phase 1: Accessibility
+- [ ] A11Y-04 — Add sr-only caption to slash commands table
+
+### Phase 2: Structural (optional)
+- [ ] DATA-01 — Extract FAQ_ENTRIES and FEATURE_GUIDES to separate data file (low priority)
+
+### Phase 3: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
