+++ specs/052-ui-audit/findings/LoginPage.md
@@ -1,0 +1,66 @@
+# Findings Report: LoginPage
+
+**Page**: LoginPage.tsx | **Route**: /login
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 119 (limit: 250) ✅ Within limit
+**Feature Dir**: src/components/auth/
+**Related Hooks**: useTheme
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 6 | 0 | 0 | 6 |
+| Loading/Error/Empty States | 5 | 0 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 6 | 1 | 0 | 7 |
+| Copy & UX Polish | 8 | 0 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 3 | 2 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **57** | **3** | **0** | **60** |
+
+**Overall**: 57/60 items passing
+
+## Findings
+
+### Minor: A11Y-03 — Decorative figure elements not aria-hidden
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-03 — ARIA attributes
+**Location**: `src/pages/LoginPage.tsx:74-107`
+**Related FR**: FR-010
+
+**Issue**: The astronaut figure constructed from multiple `<div>` elements (body, helmet, arms, etc.) is a purely decorative illustration but is not marked as aria-hidden. Screen readers may attempt to describe these empty divs.
+
+**Remediation**:
+Wrap the entire figure container in a `<div aria-hidden="true">` or add `aria-hidden="true"` to the relative container that holds all figure parts.
+
+---
+
+### Minor: TEST-01 — Login page test coverage
+
+**Dimension**: Test Coverage
+**Checklist Item**: TEST-01/02 — Tests exist
+**Location**: `src/pages/LoginPage.test.tsx`
+**Related FR**: FR-031
+
+**Issue**: LoginPage tests should cover: theme toggle button, LoginButton rendering, and accessibility attributes.
+
+---
+
+## Remediation Plan
+
+### Phase 1: Accessibility
+- [ ] A11Y-03 — Add aria-hidden="true" to decorative figure container
+
+### Phase 2: Testing
+- [ ] TEST-02 — Add accessibility tests for login page
+
+### Phase 3: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
