+++ specs/052-ui-audit/findings/NotFoundPage.md
@@ -1,0 +1,66 @@
+# Findings Report: NotFoundPage
+
+**Page**: NotFoundPage.tsx | **Route**: * (404 fallback)
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 29 (limit: 250) ✅ Well within limit
+**Feature Dir**: N/A
+**Related Hooks**: useNavigate (router only)
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 6 | 0 | 0 | 6 |
+| Loading/Error/Empty States | 5 | 0 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 6 | 1 | 0 | 7 |
+| Copy & UX Polish | 8 | 0 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 4 | 1 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **58** | **2** | **0** | **60** |
+
+**Overall**: 58/60 items passing
+
+## Findings
+
+### Minor: A11Y-03 — "404" heading text wrapped in <span> not <h2>
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-03 — ARIA attributes
+**Location**: `src/pages/NotFoundPage.tsx:10`
+**Related FR**: FR-010
+
+**Issue**: The "404" number uses a `<span>` element with text styling. While `<h1>` covers the page title, the "404" numeric indicator could be wrapped in a visually hidden context for screen readers or the span could include `aria-hidden="true"` since the `<h1>` already describes the page.
+
+**Remediation**:
+Add `aria-hidden="true"` to the "404" span since the h1 ("Lost Between Sun & Moon") already communicates the page context.
+
+---
+
+### Minor: TEST-04 — No test for 404 page rendering
+
+**Dimension**: Test Coverage
+**Checklist Item**: TEST-04 — Tests exist
+**Location**: `src/pages/NotFoundPage.test.tsx`
+**Related FR**: FR-032
+
+**Issue**: No tests exist for NotFoundPage. Should have at minimum a smoke test verifying the page renders and the "Go Home" button navigates to "/".
+
+---
+
+## Remediation Plan
+
+### Phase 1: Accessibility
+- [ ] A11Y-03 — Add aria-hidden="true" to "404" span
+
+### Phase 2: Testing
+- [ ] TEST-04 — Add smoke test for NotFoundPage
+
+### Phase 3: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
