+++ specs/052-ui-audit/findings/SettingsPage.md
@@ -1,0 +1,83 @@
+# Findings Report: SettingsPage
+
+**Page**: SettingsPage.tsx | **Route**: /settings
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 107 (limit: 250) ✅ Within limit
+**Feature Dir**: src/components/settings/
+**Related Hooks**: useUserSettings, useGlobalSettings
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 6 | 0 | 0 | 6 |
+| Loading/Error/Empty States | 4 | 1 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 6 | 1 | 0 | 7 |
+| Copy & UX Polish | 8 | 0 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 3 | 2 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **56** | **4** | **0** | **60** |
+
+**Overall**: 56/60 items passing
+
+## Findings
+
+### Major: STATE-02 — No error state when settings fail to load
+
+**Dimension**: Loading/Error/Empty States
+**Checklist Item**: STATE-02 — Error state
+**Location**: `src/pages/SettingsPage.tsx:43-52`
+**Related FR**: FR-005
+
+**Issue**: `useUserSettings` is destructured with `isLoading` and `updateSettings` but no `error` field is captured. If the settings fetch fails, the page renders with `userSettings` as `undefined`, causing the settings panels to not render (both are conditionally rendered on `userSettings &&`). The user sees a blank page with no error feedback.
+
+**Remediation**:
+Destructure `error` from `useUserSettings` and display an error banner with retry button when settings fail to load.
+
+---
+
+### Minor: A11Y-04 — Loading screen has no heading for screen readers
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-04 — Form labels
+**Location**: `src/pages/SettingsPage.tsx:67-73`
+**Related FR**: FR-011
+
+**Issue**: The loading state renders only a CelestialLoader. Screen readers would benefit from a `<h1>` or `aria-label` on the loading container to provide context.
+
+**Remediation**:
+Add `aria-label="Loading settings"` to the loading container div.
+
+---
+
+### Minor: TEST-01/TEST-02 — Settings hooks need error case tests
+
+**Dimension**: Test Coverage
+**Checklist Item**: TEST-01 — Hook tests
+**Location**: `src/hooks/useSettings.test.ts` (if exists)
+**Related FR**: FR-031
+
+**Issue**: Settings hooks may not have test coverage for error states and retry behavior.
+
+---
+
+## Remediation Plan
+
+### Phase 1: States & Error Handling
+- [ ] STATE-02 — Add error state when useUserSettings fails
+
+### Phase 2: Accessibility
+- [ ] A11Y-04 — Add aria-label to loading container
+
+### Phase 3: Testing
+- [ ] TEST-01 — Add error state test for useUserSettings
+
+### Phase 4: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
