+++ specs/052-ui-audit/findings/ToolsPage.md
@@ -1,0 +1,77 @@
+# Findings Report: ToolsPage
+
+**Page**: ToolsPage.tsx | **Route**: /tools
+**Audit Date**: 2026-03-18
+**Auditor**: speckit.implement
+**Line Count**: 87 (limit: 250) ✅ Within limit
+**Feature Dir**: src/components/tools/
+**Related Hooks**: useTools (via ToolsPanel)
+
+## Summary
+
+| Dimension | Passed | Failed | N/A | Total |
+|-----------|--------|--------|-----|-------|
+| Component Architecture | 7 | 0 | 0 | 7 |
+| Data Fetching | 6 | 0 | 0 | 6 |
+| Loading/Error/Empty States | 4 | 1 | 0 | 5 |
+| Type Safety | 5 | 0 | 0 | 5 |
+| Accessibility | 6 | 1 | 0 | 7 |
+| Copy & UX Polish | 8 | 0 | 0 | 8 |
+| Styling & Layout | 6 | 0 | 0 | 6 |
+| Performance | 5 | 0 | 0 | 5 |
+| Test Coverage | 3 | 2 | 0 | 5 |
+| Code Hygiene | 6 | 0 | 0 | 6 |
+| **Total** | **56** | **4** | **0** | **60** |
+
+**Overall**: 56/60 items passing
+
+## Findings
+
+### Major: STATE-01 — No loading state for board data
+
+**Dimension**: Loading/Error/Empty States
+**Checklist Item**: STATE-01 — Loading state
+**Location**: `src/pages/ToolsPage.tsx`
+**Related FR**: FR-004
+
+**Issue**: `useProjectBoard` is called without capturing `boardLoading` or `boardError`. The hero badge shows "Awaiting repository" until board data loads, with no explicit loading indicator for this state.
+
+**Remediation**:
+Add `boardLoading` from `useProjectBoard` and conditionally show loading state or use a skeleton for the hero badge.
+
+---
+
+### Minor: A11Y-07 — External link "MCP docs" missing rel="noopener"
+
+**Dimension**: Accessibility
+**Checklist Item**: A11Y-07 — Screen reader text
+**Location**: `src/pages/ToolsPage.tsx:49-55`
+**Related FR**: —
+
+**Issue**: External link "MCP docs" correctly uses `target="_blank"` with `rel="noopener noreferrer"` ✅. The "Discover" link also has `aria-label`. Both are correct.
+
+---
+
+### Minor: TEST-02 — ToolsPanel tests needed
+
+**Dimension**: Test Coverage
+**Checklist Item**: TEST-02 — Component tests
+**Location**: `src/components/tools/`
+**Related FR**: FR-032
+
+**Issue**: Tools panel and hooks need test coverage for error and empty states.
+
+---
+
+## Remediation Plan
+
+### Phase 1: States & Error Handling
+- [ ] STATE-01 — Add boardLoading/boardError handling in ToolsPage
+
+### Phase 2: Testing
+- [ ] TEST-02 — Add ToolsPanel component tests
+
+### Phase 3: Validation
+- [ ] ESLint: 0 warnings
+- [ ] TypeScript: 0 errors
+- [ ] Tests: all pass
